Struct stdweb::serde::Serde

pub struct Serde<T>(pub T);

A newtype which makes it possible to pass a value which implements Serde's Serializable into the js! macro.

For types defined in your crate you can also use the js_serializable! macro to make them serializable directly.

Examples

#[derive(Serialize, Debug)]
struct Person {
    name: String,
    age: i32
}

let person = Person {
   name: "Bob".to_owned(),
   age: 33
};

js! {
    var person = @{Serde( person )};
    console.log( person.name + " is " + person.age + " years old." );
};