Module stdweb::traits

A module containing reexports of all of our interface traits.

You should only import its contents through a wildcard, e.g.: use stdweb::traits::*.

Traits

IBlob
A blob object represents a file-like object of immutable, raw data. Blobs represent data that isn't necessarily in a JavaScript-native format.

IDomException
The IDomException interface represents an abnormal event which occurs as the result of calling a web API.

IElement
The IElement interface represents an object of a Document. This interface describes methods and properties common to all kinds of elements.

IError
Represents the JavaScript Error interface. An Error is thrown whenever a run-time error occurs.

IEvent
The IEvent interface represents any event which takes place in the DOM; some are user-generated (such as mouse or keyboard events), while others are generated by APIs (such as events that indicate an animation has finished running, a video has been paused, and so forth). There are many types of event, some of which use other interfaces based on the main IEvent interface. IEvent itself contains the properties and methods which are common to all events.

IEventTarget
IEventTarget is an interface implemented by objects that can receive events and may have listeners for them.

IFocusEvent
The IFocusEvent interface represents focus-related events.

IHtmlElement
The IHtmlElement interface represents any HTML element.

IKeyboardEvent
IKeyboardEvent objects describe a user interaction with the keyboard. Each event describes a key; the event type identifies what kind of activity was performed.

IMessageEvent
The MessageEvent interface represents a message received by a target object.

IMouseEvent
The IMouseEvent interface represents events that occur due to the user interacting with a pointing device (such as a mouse).

INode
INode is an interface from which a number of DOM API object types inherit.

INonElementParentNode
The INonElementParentNode mixin contains methods and properties that are common to Document and DocumentFragment.

IParentNode
The ParentNode mixin contains methods and properties that are common to all types of Node objects that can have children.

IPointerEvent
The IPointerEvent interface represents the state of a DOM event produced by a pointer such as the geometry of the contact point, the device type that generated the event, the amount of pressure that was applied on the contact surface, etc.

IProgressEvent
The IProgressEvent interface represents progress-related events.

IUiEvent
The IUiEvent interface represents simple user interface events.

IWindowOrWorker
The IWindowOrWorker mixin describes several features common to the Window and the global scope of web workers.