Module stdweb::unstable

A module containing stable counterparts to currently unstable Rust features.

Enums

Void
An uninhabited type for use in statically impossible cases.

Traits

TryFrom
Attempt to construct Self via a conversion.

TryInto
An attempted conversion that consumes self, which may or may not be expensive.