Module stdweb::web::error

A module containing error types.

Structs

DomException
A reference to a JavaScript DOMException object.

Error
A reference to a JavaScript Error object. An Error is thrown whenever a run-time error occurs.

HierarchyRequestError
Occurs when an operation would result in an incorrect node tree.

IndexSizeError
Occurs when an argument is out of range.

InvalidAccessError
Occurs when an object does not support an operation or argument.

NotFoundError
Occurs when the specified object cannot be found.

NotSupportedError
Used to indicate an unsuccessful operation when none of the other NativeError objects are an appropriate indication of the failure cause.

SecurityError
Occurs when the requested operation is insecure.

SyntaxError
Occurs when an argument does not match the expected pattern.

TypeError
Used to indicate an unsuccessful operation when none of the other NativeError objects are an appropriate indication of the failure cause.

Traits

IDomException
The IDomException interface represents an abnormal event which occurs as the result of calling a web API.

IError
Represents the JavaScript Error interface. An Error is thrown whenever a run-time error occurs.