Trait stdweb::web::event::IDragEvent

pub trait IDragEvent: IMouseEvent {
    fn data_transfer(&self) -> Option<DataTransfer> { ... }
}

The DragEvent interface is a DOM event that represents a drag and drop interaction. The user initiates a drag by placing a pointer device (such as a mouse) on the touch surface and then dragging the pointer to a new location (such as another DOM element).

This interface inherits properties from MouseEvent and Event.

(JavaScript docs)

Provided methods

fn data_transfer(&self) -> Option<DataTransfer>

The DataEvent.dataTransfer property holds the drag operation's data (as a DataTransfer object).

(JavaScript docs)

Loading content...