Trait stdweb::web::event::IKeyboardEvent

pub trait IKeyboardEvent: IEvent {
    fn alt_key(&self) -> bool { ... }

    fn code(&self) -> String { ... }

    fn ctrl_key(&self) -> bool { ... }

    fn get_modifier_state(&self, key: ModifierKey) -> bool { ... }

    fn is_composing(&self) -> bool { ... }

    fn location(&self) -> KeyboardLocation { ... }

    fn key(&self) -> String { ... }

    fn meta_key(&self) -> bool { ... }

    fn repeat(&self) -> bool { ... }

    fn shift_key(&self) -> bool { ... }
}

IKeyboardEvent objects describe a user interaction with the keyboard. Each event describes a key; the event type identifies what kind of activity was performed.

(JavaScript docs)

Provided methods

fn alt_key(&self) -> bool

Indicates whether the Alt key was down when this event was fired.

(JavaScript docs)

fn code(&self) -> String

Returns a code value that indicates the physical key pressed on the keyboard.

(JavaScript docs)

fn ctrl_key(&self) -> bool

Returns whether the Ctrl key was down when this event was fired.

(JavaScript docs)

fn get_modifier_state(&self, key: ModifierKey) -> bool

Returns whether a modifier key was down when this event was fired.

(JavaScript docs)

fn is_composing(&self) -> bool

Returns whether this event was fired during composition.

(JavaScript docs)

fn location(&self) -> KeyboardLocation

Returns the location of the key on the keyboard.

(JavaScript docs)

fn key(&self) -> String

Returns the value of a key or keys pressed by the user.

(JavaScript docs)

fn meta_key(&self) -> bool

Indicates whether the Meta key was down when this event was fired.

(JavaScript docs)

fn repeat(&self) -> bool

Indicates whether the key is held down such that it is repeating.

(JavaScript docs)

fn shift_key(&self) -> bool

Indicates whether the Shift key was down when this event was fired.

(JavaScript docs)

Loading content...