Trait stdweb::web::event::IMessageEvent

pub trait IMessageEvent: IEvent where
    <Self::Data as TryFrom<Value>>::Error: Debug, {
    type Data: TryFrom<Value>;
    fn data(&self) -> Self::Data { ... }

    fn origin(&self) -> String { ... }

    fn last_event_id(&self) -> String { ... }

    fn source(&self) -> Option<Reference> { ... }

    fn ports(&self) -> Vec<Reference> { ... }
}

The MessageEvent interface represents a message received by a target object.

(JavaScript docs)

Associated Types

type Data: TryFrom<Value>

The type of data received with this MessageEvent

Loading content...
Provided methods

fn data(&self) -> Self::Data

The data sent by the message emitter.

(JavaScript docs)

fn origin(&self) -> String

A string representing the origin of the message emitter.

(JavaScript docs)

fn last_event_id(&self) -> String

A string representing a unique ID for the event.

(JavaScript docs)

fn source(&self) -> Option<Reference>

A MessageEventSource (which can be a WindowProxy, MessagePort, or ServiceWorker object) representing the message emitter.

(JavaScript docs)

fn ports(&self) -> Vec<Reference>

An array of MessagePort objects representing the ports associated with the channel the message is being sent through (where appropriate, e.g. in channel messaging or when sending a message to a shared worker).

(JavaScript docs)

Loading content...