Trait stdweb::web::event::IProgressEvent

pub trait IProgressEvent: IEvent {
    fn length_computable(&self) -> bool { ... }

    fn loaded(&self) -> u64 { ... }

    fn total(&self) -> u64 { ... }
}

The IProgressEvent interface represents progress-related events.

(JavaScript docs)

Provided methods

fn length_computable(&self) -> bool

Indicates whether the progress is measureable.

(JavaScript docs)

fn loaded(&self) -> u64

Returns the amount of work already performed by the underlying process.

(JavaScript docs)

fn total(&self) -> u64

Returns the total amount of work that the underlying process will perform.

(JavaScript docs)

Loading content...