Struct stdweb::web::html_element::CanvasElement

pub struct CanvasElement(_);

The HTML <canvas> element provides an empty graphic zone on which specific JavaScript APIs can draw (such as Canvas 2D or WebGL).

(JavaScript docs)

Methods

impl CanvasElement

pub fn height(&self) -> u32

Returns a positive integer reflecting the height HTML attribute of the element interpreted in CSS pixels. When the attribute is not specified, or if it is set to an invalid value, like a negative, the default value of 150 is used.

(JavaScript docs)

pub fn set_height(&self, value: u32)

Sets a positive integer reflecting the height HTML attribute of the element interpreted in CSS pixels. When the attribute is not specified, or if it is set to an invalid value, like a negative, the default value of 150 is used.

(JavaScript docs)

pub fn width(&self) -> u32

Returns a positive integer reflecting the width HTML attribute of the element interpreted in CSS pixels. When the attribute is not specified, or if it is set to an invalid value, like a negative, the default value of 300 is used.

(JavaScript docs)

pub fn set_width(&self, value: u32)

Sets a positive integer reflecting the width HTML attribute of the element interpreted in CSS pixels. When the attribute is not specified, or if it is set to an invalid value, like a negative, the default value of 300 is used.

(JavaScript docs)

pub fn get_context<T: RenderingContext>(&self) -> Result<T, T::Error>

Returns a drawing context on the canvas, or None if the context identifier is not supported.

(JavaScript docs)

pub fn to_data_url(
&self,
mime_type: Option<&str>,
quality: Option<f64>
) -> Result<String, TODO>

Returns a data URI containing a representation of the image in the format specified by the type parameter (defaults to PNG). The returned image is in a resolution of 96 dpi.

(JavaScript docs)

pub fn to_blob<F: FnOnce(Blob) + 'static>(
&self,
f: F,
mime_type: Option<&str>,
quality: Option<f64>
) -> Result<(), TODO>

Creates a Blob object representing the image contained in the canvas; this file may be cached on the disk or stored in memory at the discretion of the user agent.

(JavaScript docs)