Struct stdweb::web::CanvasGradient

pub struct CanvasGradient(_);

The CanvasGradient struct represents an opaque object describing a gradient. It is returned by the methods CanvasRenderingContext2D.createLinearGradient() or CanvasRenderingContext2D.createRadialGradient().

(JavaScript docs)

Methods

impl CanvasGradient

pub fn add_color_stop(
&self,
offset: f64,
color: &str
) -> Result<(), AddColorStopError>

Adds a new stop, defined by an offset and a color, to the gradient. If the offset is not between 0 and 1, an INDEX_SIZE_ERR is returned, if the color can't be parsed as a CSS , a SYNTAX_ERR is returned.

(JavaScript docs)