Struct stdweb::web::History

pub struct History(_);

(JavaScript docs)

Methods

impl History

pub fn push_state<T: JsSerialize>(
&self,
state: T,
title: &str,
url: Option<&str>
)

Adds a new entry to history.

pushState() takes three parameters: a state object, a title (which is currently ignored), and (optionally) a URL. Let's examine each of these three parameters in more detail:

state object — The state object is a JavaScript object which is associated with the new history entry created by pushState(). Whenever the user navigates to the new state, a popstate event is fired, and the state property of the event contains a copy of the history entry's state object.

title — Firefox currently ignores this parameter, although it may use it in the future. Passing the empty string here should be safe against future changes to the method. Alternatively, you could pass a short title for the state to which you're moving.

URL — The new history entry's URL is given by this parameter. Note that the browser won't attempt to load this URL after a call to pushState(), but it might attempt to load the URL later, for instance after the user restarts the browser. The new URL does not need to be absolute; if it's relative, it's resolved relative to the current URL. The new URL must be of the same origin as the current URL; otherwise, pushState() will throw an exception. This parameter is optional; if it isn't specified, it's set to the document's current URL.

(JavaScript docs)

pub fn replace_state<T: JsSerialize>(
&self,
state: T,
title: &str,
url: Option<&str>
) -> Result<(), TODO>

Operates exactly like history.push_state() except that replace_state() modifies the current history entry instead of creating a new one. Note that this doesn't prevent the creation of a new entry in the global browser history.

(JavaScript docs)

pub fn go(&self, offset: i32) -> Result<(), TODO>

You can use the go() method to load a specific page from session history, identified by its relative position to the current page (with the current page being, of course, relative index 0).

(JavaScript docs)

pub fn back(&self) -> Result<(), TODO>

Move one step backward through history.

(JavaScript docs)

pub fn forward(&self) -> Result<(), TODO>

Move one step forward through history.

(JavaScript docs)

pub fn len(&self) -> u32

Returns the current number of history entries.

(JavaScript docs)