Struct stdweb::web::Location

pub struct Location(_);

The Location interface represents the location (URL) of the object it is linked to. Changes done on it are reflected on the object it relates to. Both the Document and Window interface have such a linked Location, accessible via Document::location and Window::location respectively.

Note that all Location methods can return a SecurityError if the Location object's relevant Document's origin is not same origin-domain with the entry settings object's origin. See: https://html.spec.whatwg.org/#dom-location-href

(JavaScript docs)

Methods

impl Location

pub fn href(&self) -> Result<String, SecurityError>

The entire URL.

(JavaScript docs)

pub fn origin(&self) -> Result<String, SecurityError>

Returns a String containing the Unicode serialization of the origin of the represented URL, that is:

For URL using the http or https, the scheme followed by '://', followed by the domain, followed by ':', followed by the port (the default port, 80 and 443 respectively, if explicitely specified);

For URL using file: scheme, the value is browser dependant.

For URL using the blob: scheme, the origin of the URL following blob:. E.g "blob:https://mozilla.org" will have "https://mozilla.org".

(JavaScript docs)

pub fn protocol(&self) -> Result<String, SecurityError>

Returns a String representing the protocol scheme of the URL, including the final ':'.

Example: http:

(JavaScript docs)

pub fn host(&self) -> Result<String, SecurityError>

Returns a String containing the host (i.e. hostname) and then, if the port of the URL is nonempty, a ':', and the port of the URL.

Example: hitchhikers.com:4242

(JavaScript docs)

pub fn hostname(&self) -> Result<String, SecurityError>

Returns a String which is the domain of the URL

Example: mozilla.com

(JavaScript docs)

pub fn port(&self) -> Result<String, SecurityError>

Returns a String containing the port number or "" if there is no port.

(JavaScript docs)

pub fn pathname(&self) -> Result<String, SecurityError>

Returns a String containing an initial '/' followed by the path of the URL.

(JavaScript docs)

pub fn search(&self) -> Result<String, SecurityError>

Returns a String which is a search string, also called a query string, that is a String containing a '?' followed by the parameters of the URL.

These can then be further parsed via another library.

(JavaScript docs)

pub fn hash(&self) -> Result<String, SecurityError>

Returns a String containing a '#' followed by the fragment identifier of the URL. The fragment is not percent-decoded.

(JavaScript docs)