Struct stdweb::web::MutationObserver

pub struct MutationObserver(_);

Provides a way to receive notifications about changes to the DOM.

(JavaScript docs)

Methods

impl MutationObserver

pub fn new<F>(callback: F) -> MutationObserverHandle where
F: FnMut(Vec<MutationRecord>, Self) + 'static,

Returns a new MutationObserverHandle with the given callback.

The callback will be called with the following arguments when the observed DOM nodes change:

A vector of changes to the observed DOM nodes.

The MutationObserver.

(JavaScript docs)

pub fn observe<T: INode>(
&self,
target: &T,
options: MutationObserverInit
) -> Result<(), TODO>

Starts observing changes to the target.

When the target is changed, the MutationObserver is notified with a vector of MutationRecord.

The options specifies which changes should be observed.

Multiple different targets can be observed simultaneously (with the same or different options).

If you call observe on the same target multiple times, it will replace the old options with the new options. It will not notify multiple times for the same change to the same target.

Panics

At least one of child_list, attributes, or character_data must be true.

If attribute_old_value is true, then attributes must be true.

If character_data_old_value is true, then character_data must be true.

If attribute_filter is Some, then attributes must be true.

(JavaScript docs)

pub fn disconnect(&self)

Stops observing all targets.

Until the observe method is called again, the MutationObserver will not be notified of any changes.

(JavaScript docs)

pub fn take_records(&self) -> Vec<MutationRecord>

Empties the MutationObserver's record queue and returns what was in there.

This method is generally not needed, instead use the MutationObserver callback to respond to changes.

(JavaScript docs)