Struct stdweb::web::MutationObserverHandle

pub struct MutationObserverHandle { /* fields omitted */ }

A wrapper which ensures that memory is properly cleaned up when it's no longer needed.

This is created by the MutationObserver::new method, and it can use the same methods as MutationObserver.

When the MutationObserverHandle is dropped, the disconnect method will automatically be called.