Struct stdweb::web::MutationObserverInit

pub struct MutationObserverInit<'a> {
    pub child_list: bool,
    pub attributes: bool,
    pub character_data: bool,
    pub subtree: bool,
    pub attribute_old_value: bool,
    pub character_data_old_value: bool,
    pub attribute_filter: Option<&'a [&'a str]>,
}

Specifies which changes should be observed for the target.

This is only used with the MutationObserver::observe method.

(JavaScript docs)

Fields
child_list: bool

If true it will observe all inserts and removals of the target's children (including text nodes).

This is not recursive, it will only observe immediate children (unless subtree is true in which case it will observe all children and sub-children recursively).

attributes: bool

If true it will observe all changes to the target's attributes.

character_data: bool

If true it will observe all changes to the CharacterData's data.

subtree: bool

If true it will observe all changes to the target, the target's children, and the target's sub-children.

This is recursive, so it causes all children and sub-children to be observed.

attribute_old_value: bool

If true it will store the target's old attribute value in old_value.

character_data_old_value: bool

If true it will store the CharacterData's old data in old_data.

attribute_filter: Option<&'a [&'a str]>

If Some it will only observe the specified attributes. The attributes should be specified without a namespace.

If None it will observe all attributes.