Struct stdweb::web::NodeList

pub struct NodeList(_);

NodeList objects are collections of nodes such as those returned by properties such as INode::child_nodes and the Document::query_selector_all method.

In some cases, the NodeList is a live collection, which means that changes in the DOM are reflected in the collection - for example INode::child_nodes is live.

In other cases, the NodeList is a static collection, meaning any subsequent change in the DOM does not affect the content of the collection - for example Document::query_selector_all returns a static NodeList.

(JavaScript docs)

Methods

impl NodeList

pub fn len(&self) -> u32

Returns the number of Nodes contained in this list.

(JavaScript docs)

pub fn iter(&self) -> NodeIter

Returns an iterator over the list.