Struct stdweb::web::SocketCloseCode

pub struct SocketCloseCode(pub u16);

Wrapper type around a CloseEvent code, indicating why the WebSocket was closed

(JavaScript docs)

Methods

impl SocketCloseCode

pub const NORMAL_CLOSURE: SocketCloseCode

NORMAL_CLOSURE: SocketCloseCode = SocketCloseCode(1000)

Normal closure; the connection successfully completed whatever purpose for which it was created.

pub const GOING_AWAY: SocketCloseCode

GOING_AWAY: SocketCloseCode = SocketCloseCode(1001)

The endpoint is going away, either because of a server failure or because the browser is navigating away from the page that opened the connection.

pub const PROTOCOL_ERROR: SocketCloseCode

PROTOCOL_ERROR: SocketCloseCode = SocketCloseCode(1002)

The endpoint is terminating the connection due to a protocol error.

pub const UNSUPPORTED_DATA: SocketCloseCode

UNSUPPORTED_DATA: SocketCloseCode = SocketCloseCode(1003)

The connection is being terminated because the endpoint received data of a type it cannot accept (for example, a text-only endpoint received binary data).

pub const NO_STATUS_RECEIVED: SocketCloseCode

NO_STATUS_RECEIVED: SocketCloseCode = SocketCloseCode(1005)

Reserved. Indicates that no status code was provided even though one was expected.

pub const ABNORMAL_CLOSURE: SocketCloseCode

ABNORMAL_CLOSURE: SocketCloseCode = SocketCloseCode(1006)

Reserved. Used to indicate that a connection was closed abnormally (that is, with no close frame being sent) when a status code is expected.

pub const INVALID_FRAME_PAYLOAD_DATA: SocketCloseCode

INVALID_FRAME_PAYLOAD_DATA: SocketCloseCode = SocketCloseCode(1007)

The endpoint is terminating the connection because a message was received that contained inconsistent data (e.g., non-UTF-8 data within a text message).

pub const POLICY_VIOLATION: SocketCloseCode

POLICY_VIOLATION: SocketCloseCode = SocketCloseCode(1008)

The endpoint is terminating the connection because it received a message that violates its policy. This is a generic status code, used when codes 1003 and 1009 are not suitable.

pub const MESSAGE_TOO_BIG: SocketCloseCode

MESSAGE_TOO_BIG: SocketCloseCode = SocketCloseCode(1009)

The endpoint is terminating the connection because a data frame was received that is too large.

pub const MISSING_EXTENSION: SocketCloseCode

MISSING_EXTENSION: SocketCloseCode = SocketCloseCode(1010)

The client is terminating the connection because it expected the server to negotiate one or more extensions, but the server didn't.

pub const INTERNAL_ERROR: SocketCloseCode

INTERNAL_ERROR: SocketCloseCode = SocketCloseCode(1011)

The server is terminating the connection because it encountered an unexpected condition that prevented it from fulfilling the request.

pub const SERVICE_RESTART: SocketCloseCode

SERVICE_RESTART: SocketCloseCode = SocketCloseCode(1012)

The server is terminating the connection because it is restarting.

pub const TRY_AGAIN_LATER: SocketCloseCode

TRY_AGAIN_LATER: SocketCloseCode = SocketCloseCode(1013)

The server is terminating the connection due to a temporary condition, e.g. it is overloaded and is casting off some of its clients.

pub const BAD_GATEWAY: SocketCloseCode

BAD_GATEWAY: SocketCloseCode = SocketCloseCode(1014)

The server was acting as a gateway or proxy and received an invalid response from the upstream server. This is similar to 502 HTTP Status Code.

pub const TLS_HANDSHAKE: SocketCloseCode

TLS_HANDSHAKE: SocketCloseCode = SocketCloseCode(1015)

Reserved. Indicates that the connection was closed due to a failure to perform a TLS handshake (e.g., the server certificate can't be verified).