Struct stdweb::web::Storage

pub struct Storage(_);

The Storage interface of the Web Storage API provides access to the session storage or local storage for a particular domain.

(JavaScript docs)

Methods

impl Storage

pub fn len(&self) -> u32

Gets the number of data items stored in the Storage object.

(JavaScript docs)

pub fn get(&self, key: &str) -> Option<String>

Returns a value corresponding to the key.

(JavaScript docs)

pub fn insert(&self, key: &str, value: &str) -> Result<(), TODO>

Inserts a key-value pair into the storage.

(JavaScript docs)

pub fn remove(&self, key: &str)

Removes a key from the storage.

(JavaScript docs)

pub fn clear(&self)

When invoked, will empty all keys out of the storage.

(JavaScript docs)

pub fn key(&self, nth: u32) -> Option<String>

Return the name of the nth key in the storage.

(JavaScript docs)

pub fn contains_key(&self, key: &str) -> bool

Returns true if the storage contains a value for the specified key.