Struct stdweb::web::WebSocket

pub struct WebSocket(_);

The WebSocket object provides the API for creating and managing a WebSocket connection to a server, as well as for sending and receiving data on the connection.

(JavaScript docs)

Methods

impl WebSocket

pub fn new(url: &str) -> Result<WebSocket, CreationError>

Returns a newly constructed WebSocket.

(JavaScript docs)

pub fn new_with_protocols(
url: &str,
protocols: &[&str]
) -> Result<WebSocket, CreationError>

Returns a newly constructed WebSocket.

(JavaScript docs)

pub fn binary_type(&self) -> SocketBinaryType

Returns the binary type of the web socket. Only affects received messages. The default binary type is Blob.

(JavaScript docs)

pub fn set_binary_type(&self, binary_type: SocketBinaryType)

Sets the binary type of the web socket. Only affects received messages. The default binary type is Blob.

(JavaScript docs)

pub fn buffered_amount(&self) -> u64

Returns the number of bytes of data that have been queued using calls to send() but not yet transmitted to the network. This value resets to zero once all queued data has been sent. This value does not reset to zero when the connection is closed; if you keep calling send(), this will continue to climb.

(JavaScript docs)

pub fn extensions(&self) -> String

Returns the extensions selected by the server. This is currently only the empty string or a list of extensions as negotiated by the connection.

(JavaScript docs)

pub fn protocol(&self) -> String

Returns a string indicating the name of the sub-protocol the server selected; this will be one of the strings specified in the protocols parameter when creating the WebSocket object.

(JavaScript docs)

pub fn ready_state(&self) -> SocketReadyState

Returns the current state of the connection.

(JavaScript docs)

pub fn url(&self) -> String

Returns the URL as resolved by the constructor. This is always an absolute URL.

(JavaScript docs)

pub fn close(&self)

Closes the WebSocket connection or connection attempt, if any. If the connection is already CLOSED, this method does nothing.

(JavaScript docs)

pub fn close_with_status(
&self,
code: SocketCloseCode,
reason: &str
) -> Result<(), CloseError>

Closes the WebSocket connection or connection attempt, if any. If the connection is already CLOSED, this method does nothing.

(JavaScript docs)

pub fn send_text(&self, text: &str) -> Result<(), TODO>

Enqueues the specified data to be transmitted to the server over the WebSocket connection, increasing the value of bufferedAmount by the number of bytes needed to contain the data. If the data can't be sent (for example, because it needs to be buffered but the buffer is full), the socket is closed automatically.

pub fn send_blob(&self, blob: &Blob) -> Result<(), TODO>

Enqueues the specified data to be transmitted to the server over the WebSocket connection, increasing the value of bufferedAmount by the number of bytes needed to contain the data. If the data can't be sent (for example, because it needs to be buffered but the buffer is full), the socket is closed automatically.

pub fn send_array_buffer(&self, array_buffer: &ArrayBuffer) -> Result<(), TODO>

Enqueues the specified data to be transmitted to the server over the WebSocket connection, increasing the value of bufferedAmount by the number of bytes needed to contain the data. If the data can't be sent (for example, because it needs to be buffered but the buffer is full), the socket is closed automatically.

pub fn send_bytes(&self, bytes: &[u8]) -> Result<(), TODO>

Enqueues the specified data to be transmitted to the server over the WebSocket connection, increasing the value of bufferedAmount by the number of bytes needed to contain the data. If the data can't be sent (for example, because it needs to be buffered but the buffer is full), the socket is closed automatically.