Struct stdweb::web::XmlHttpRequest

pub struct XmlHttpRequest(_);

Use XmlHttpRequest (XHR) objects to interact with servers. You can retrieve data from a URL without having to do a full page refresh. This enables a Web page to update just part of a page without disrupting what the user is doing. XmlHttpRequest is used heavily in Ajax programming.

(JavaScript docs)

Methods

impl XmlHttpRequest

pub fn new() -> XmlHttpRequest

Creates new XmlHttpRequest.

pub fn ready_state(&self) -> XhrReadyState

Returns the current state of the request as a XhrReadyState.

(JavaScript docs)

pub fn response_text(&self) -> Result<Option<String>, TODO>

Returns a string that contains the response to the request as text, or None if the request was unsuccessful or has not yet been sent.

(JavaScript docs)

pub fn status(&self) -> u16

Returns an unsigned short with the status of the response of the request.

(JavaScript docs)

pub fn open(&self, method: &str, url: &str) -> Result<(), TODO>

Open connection with given method (ie GET or POST), and the url to hit.

(JavaScript docs)

pub fn get_response_header(&self, header: &str) -> Option<String>

Returns the string containing the text of the specified header. If there are multiple response headers with the same name, then their values are returned as a single concatenated string.

(JavaScript docs)

pub fn set_request_header(&self, header: &str, value: &str) -> Result<(), TODO>

Sets the value of an HTTP request header. Must be called after open(), but before send(). If this method is called several times with the same header, the values are merged into one single request header.

(JavaScript docs)

pub fn send(&self) -> Result<(), TODO>

Send request on an open connection with no data

(JavaScript docs)

pub fn send_with_string(&self, body: &str) -> Result<(), TODO>

Send request on an open connection with string body

(JavaScript docs)

pub fn send_with_bytes(&self, body: &[u8]) -> Result<(), TODO>

Send request on an open connection with a byte array body

(JavaScript docs)

pub fn abort(&self)

Aborts the request if it has already been sent. When a request is aborted, its ready_state is changed to Done and the status code is set to Unsent.

(JavaScript docs)