Trait stdweb::web::IEventTarget

pub trait IEventTarget: ReferenceType {
    fn add_event_listener<T, F>(&self, listener: F) -> EventListenerHandle
    where
        T: ConcreteEvent,
        F: FnMut(T) + 'static,
    { ... }

    fn dispatch_event<T: IEvent>(&self, event: &T) -> Result<bool, TODO> { ... }
}

IEventTarget is an interface implemented by objects that can receive events and may have listeners for them.

(JavaScript docs)

Provided methods

fn add_event_listener<T, F>(&self, listener: F) -> EventListenerHandle where
T: ConcreteEvent,
F: FnMut(T) + 'static,

Adds given event handler to the list of event listeners for the specified EventTarget on which it's called.

(JavaScript docs)

fn dispatch_event<T: IEvent>(&self, event: &T) -> Result<bool, TODO>

Dispatches an Event at this EventTarget, invoking the affected event listeners in the appropriate order.

(JavaScript docs)

Loading content...