Trait stdweb::web::INonElementParentNode

pub trait INonElementParentNode: ReferenceType {
    fn get_element_by_id(&self, id: &str) -> Option<Element> { ... }
}

The INonElementParentNode mixin contains methods and properties that are common to Document and DocumentFragment.

You most likely don't want to use this directly; instead you should use stdweb::traits::*;.

Provided methods

fn get_element_by_id(&self, id: &str) -> Option<Element>

Returns a reference to the element by its ID; the ID is a string which can be used to uniquely identify the element, found in the HTML id attribute.

(JavaScript docs)

Loading content...