Enum stdweb::Value

pub enum Value {
    Undefined,
    Null,
    Bool(bool),
    Number(Number),
    Symbol(Symbol),
    String(String),
    Reference(Reference),
}

A type representing a JavaScript value.

This type implements a rich set of conversions from and into standard Rust types, for example:

let v1: Value = "Hello world!".into();
let v2: Value = true.into();
let v3: Value = vec![ 1, 2, 3 ].into();
let v4: Value = Null.into();
let v5: Value = 123_u64.try_into().unwrap();

let v1_r: String = v1.try_into().unwrap();
let v2_r: bool = v2.try_into().unwrap();
let v3_r: Vec< i32 > = v3.try_into().unwrap();
let v4_r: Option< String > = v4.try_into().unwrap(); // Will be `None`.
let v5_r: u64 = v5.try_into().unwrap();

Variants

Undefined

Null

Bool(bool)

Number(Number)

Symbol(Symbol)

String(String)

Reference(Reference)

Methods

impl Value

pub fn is_null(&self) -> bool

Checks whenever the Value is of the Null variant.

pub fn is_symbol(&self) -> bool

Checks whenever the Value is of the Symbol variant.

pub fn is_reference(&self) -> bool

Checks whenever the Value is of the Reference variant.

pub fn is_object(&self) -> bool

Checks whenever the Value is a reference to an Object.

pub fn is_array(&self) -> bool

Checks whenever the Value is a reference to an Array.

pub fn as_reference(&self) -> Option<&Reference>

Gets a reference to the Reference inside this Value.

pub fn as_object(&self) -> Option<&Object>

Gets a reference to the Object inside this Value.

pub fn as_array(&self) -> Option<&Array>

Gets a reference to the Array inside this Value.

pub fn into_reference(self) -> Option<Reference>

Returns the Reference inside this Value.

pub fn into_object(self) -> Option<Object>

Returns the Object inside this Value.

pub fn into_array(self) -> Option<Array>

Returns the Array inside this Value.

pub unsafe fn into_reference_unchecked<T: ReferenceType>(self) -> Option<T>

Converts a Reference inside this Value into the given type T; doesn't check whenever the reference is really of type T.

In cases where the value is not a Reference a None is returned.

pub fn into_string(self) -> Option<String>

Returns the String inside this Value.

pub fn as_str(&self) -> Option<&str>

Returns a borrow of the string inside this Value.