Trait stdweb::InstanceOf

pub trait InstanceOf {
    fn instance_of(reference: &Reference) -> bool;
}

A trait to check whenever a given Reference is of a certain type.

Required methods

fn instance_of(reference: &Reference) -> bool

Checks whenever a given Reference if of type Self.

Loading content...