Trait stdweb::JsSerialize

pub trait JsSerialize { }

A trait for types which can be serialized through the js! macro.

Do not try to implement this trait yourself! It's only meant to be used inside generic code for specifying trait bounds.