Module stdweb::web::error

A module containing error types.

Structs

AbortError
Used to indicate that the operation was aborted.

DomException
A reference to a JavaScript DOMException object.

Error
A reference to a JavaScript Error object. An Error is thrown whenever a run-time error occurs.

HierarchyRequestError
Occurs when an operation would result in an incorrect node tree.

IndexSizeError
Occurs when an argument is out of range.

InvalidAccessError
Occurs when an object does not support an operation or argument.

InvalidCharacterError
Used to indicate the string contains one or more characters which are invalid.

InvalidStateError
Occurs when an object is in an invalid state.

NotFoundError
Occurs when the specified object cannot be found.

NotSupportedError
Used to indicate an unsuccessful operation when none of the other NativeError objects are an appropriate indication of the failure cause.

SecurityError
Occurs when the requested operation is insecure.

SyntaxError
Occurs when an argument does not match the expected pattern.

TypeError
Used to indicate an unsuccessful operation when none of the other NativeError objects are an appropriate indication of the failure cause.

UnknownValueError
Indicates that an invalid value is setted to an SelectElement. It means there is no <option> element that has the given value.

Enums

AddColorStopError
A enum of the exceptions that CanvasGradient.add_color_stop() may throw

DrawImageError
A enum of the exceptions that CanvasRenderingContext2D.draw_image() and similar may throw

GetImageDataError
A enum of the exceptions that CanvasRenderingContext2D.get_image_data() may throw

XhrSetResponseTypeError
An error returned from XmlHttpRequest::set_response_type

Traits

IDomException
The IDomException interface represents an abnormal event which occurs as the result of calling a web API.

IError
Represents the JavaScript Error interface. An Error is thrown whenever a run-time error occurs.