Module stdweb::web::html_element

A module containing HTML DOM elements.

Structs

CanvasElement
The HTML <canvas> element provides an empty graphic zone on which specific JavaScript APIs can draw (such as Canvas 2D or WebGL).

ImageElement
The HTML image element is used to manipulate the layout and presentation of <img> elements.

InputElement
The HTML input element is used to create interactive controls for web-based forms in order to accept data from the user.

OptionElement
The HTML <option> element is used to define an item contained in a <select>, an <optgroup>, or a <datalist> element.

SelectElement
The HTML <select> element represents a control that provides a menu of options.

SlotElement
The HTML <slot> element represents a placeholder inside a web component that you can fill with your own markup, which lets you create separate DOM trees and present them together.

TemplateElement
The HTML <template> element represents a mechanism for holding client-side content that is not to be rendered when a page is loaded but may subsequently be instantiated during runtime using JavaScript.

TextAreaElement
The HTML <textarea> element represents a multi-line plain-text editing control.