Struct stdweb::web::Window

pub struct Window(_);

The Window object represents a window containing a DOM document.

(JavaScript docs)

Methods

impl Window

pub fn alert(&self, message: &str)

The Window.alert() method displays an alert dialog with the optional specified content and an OK button.

(JavaScript docs)

pub fn confirm(&self, message: &str) -> bool

The Window.confirm() method displays a modal dialog with an optional message and two buttons: OK and Cancel.

(JavaScript docs)

pub fn local_storage(&self) -> Storage

The local_storage property allows you to access a local Storage object.

It is similar to the Window::session_storage. The only difference is that, while data stored in local_storage has no expiration time, data stored in session_storage gets cleared when the browsing session ends - that is, when the browser is closed.

(JavaScript docs)

pub fn session_storage(&self) -> Storage

The session_storage property allows you to access a session Storage object for the current origin.

It is similar to the Window::local_storage, The only difference is that, while data stored in local_storage has no expiration time, data stored in session_storage gets cleared when the browsing session ends.

A page session lasts for as long as the browser is open and survives over page reloads and restores. Opening a page in a new tab or window will cause a new session to be initiated, which differs from how session cookies work.

(JavaScript docs)

pub fn location(&self) -> Option<Location>

Returns a Location object which contains information about the URL of the document and provides methods for changing that URL and loading another URL.

(JavaScript docs)

pub fn request_animation_frame<F: FnOnce(f64) + 'static>(
&self,
callback: F
) -> RequestAnimationFrameHandle

You should call this method whenever you're ready to update your animation onscreen. This will request that your animation function be called before the browser performs the next repaint. The number of callbacks is usually 60 times per second, but will generally match the display refresh rate in most web browsers as per W3C recommendation. request_animation_frame() calls are paused in most browsers when running in background tabs or hidden iframes in order to improve performance and battery life.

The callback method is passed a single argument, a f64, which indicates the current time when callbacks queued by requestAnimationFrame() begin to fire. Multiple callbacks in a single frame, therefore, each receive the same timestamp even though time has passed during the computation of every previous callback's workload. This timestamp is a decimal number, in milliseconds, but with a minimal precision of 1ms (1000 µs).

(JavaScript docs)

pub fn history(&self) -> History

Returns the global History object, which provides methods to manipulate the browser history.

(JavaScript docs)

pub fn inner_width(&self) -> i32

Returns the width (in pixels) of the browser window viewport including, if rendered, the vertical scrollbar.

(JavaScript docs)

pub fn inner_height(&self) -> i32

Returns the height (in pixels) of the browser window viewport including, if rendered, the horizontal scrollbar.

(JavaScript docs)

pub fn outer_width(&self) -> i32

Returns the width of the outside of the browser window. It represents the width of the whole browser window including sidebar (if expanded), window chrome and window resizing borders/handles.

(JavaScript docs)

pub fn outer_height(&self) -> i32

Returns the height of the outside of the browser window. It represents the height of the whole browser window including sidebar (if expanded), window chrome and window resizing borders/handles.

(JavaScript docs)

pub fn page_y_offset(&self) -> f64

The read-only Window property pageYOffset is an alias for scrollY; as such, it returns the number of pixels the document is currently scrolled along the vertical axis (that is, up or down), with a value of 0.0 indicating that the top edge of the Document is currently aligned with the top edge of the window's content area.

(JavaScript docs)

pub fn page_x_offset(&self) -> f64

This is an alias for scrollX.

(JavaScript docs)

pub fn device_pixel_ratio(&self) -> f64

The ratio in resolution from physical pixels to CSS pixels

(JavaScript docs)

pub fn get_selection(&self) -> Option<Selection>

Returns a Selection object representing the range of text selected by the user or the current position of the caret. (Javascript docs)